+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.6.8
+- New objective handling: Introduce subobjective classes for more intuitive parameterization of weights. As well as for reading out the solved subobjective functions.
+- Update Dashboard to visualize subobjectives
+
 ## 0.6.7
 - #60: Forbid users from creating instance attributes which are also variables in CasadiModel